plots: always start from the right #457

+++ include/notcurses/notcurses.h
@@ -2557,7 +2557,7 @@ typedef struct ncplot_options {
   // of keys. for a time range, say the previous hour sampled with second
   // resolution, the independent variable would be the range [0..3600): 3600.
   // if rangex is 0, it is dynamically set to the number of columns.
-  uint64_t rangex;
+  int rangex;
   uint64_t miny, maxy; // y axis min and max. for autodiscovery, set them equal.
   bool labelaxisd; // generate labels for the dependent axis
   bool exponentialy;  // is y-axis exponential? (not yet implemented)
